# Use handwritten cosine similarity